+++ src/tune.cpp
@@ -100,10 +100,10 @@ template<> void Tune::Entry<Score>::init_option() {
 
 template<> void Tune::Entry<Score>::read_option() {
   if (Options.count("m" + name))
-      value = make_score(static_cast<int>(Options["m" + name]), eg_value(value));
+      value = make_score(Options["m" + name], eg_value(value));
 
   if (Options.count("e" + name))
-      value = make_score(mg_value(value), static_cast<int>(Options["e" + name]));
+      value = make_score(mg_value(value), Options["e" + name]);
 }
 
 // Instead of a variable here we have a PostUpdate function: just call it
